requirements.txt
langchain==0.1.12
langchain-community==0.0.27
openai==1.14.0
python-dotenv==1.0.1
pypdf==4.1.0
yt-dlp==2024.3.10
pydub==0.25.1 